+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- Add spell base power (`base_power`) to parser, schema and models.
+- Add `slot` and `base_power` to item attributes.
+
 ## 8.0.1 (2026-06-16)
 
 - Add support for new NPC transport template.